P6altneu/requirements.txt
#logging
mysql-connector-python
#argparse
#black
#datetime
